Move resolve.ts into service/index.ts
