+++ src/common/index.ts
@@ -9,4 +9,4 @@
 export { default as CONSTANT } from './constant';
 export { resolveReactRoutes } from './routes';
 export { getStatusCode } from './status-codes';
-export { createMessage, IPCRequestObject, IPCResponseObject } from './util';
+export { createMessage } from './util';
