+++ src/common/util.ts
@@ -19,49 +19,6 @@ declare global {
 }
 
 /**
- * Returns IPC Request Object to be sent to main process.
- *
- * @param URI URI String.
- * @param data Payload.
- * @returns {object}
- */
-const IPCRequestObject = (
-  URI: string,
-  data?: unknown | object
-): IPCRequestInterface => {
-  return {
-    URI,
-    timestamp: Date.now(),
-    data,
-  };
-};
-
-/**
- * Returns IPC Response Object to be sent to renderer process.
- *
- * @param URI URI String received from renderer process.
- * @param timestamp Timestamp of  renderer's IPCRequestObject.
- * @param messageObject Request Fulfillment status and message.
- * @param data Payload.
- * @returns {object}
- */
-const IPCResponseObject = (
-  URI: string,
-  timestamp: number,
-  messageObject: MessageInterface,
-  data?: unknown | object
-): IPCResponseInterface => {
-  const { status, message } = messageObject;
-  return {
-    URI,
-    timestamp,
-    status,
-    data,
-    message,
-  };
-};
-
-/**
  * Return object implementing MessageInterface.
  *
  * @param status Status Code.
@@ -76,4 +33,4 @@ const createMessage = (
   message,
 });
 
-export { createMessage, IPCRequestObject, IPCResponseObject };
+export { createMessage };
